Add observable remember
